Removing compiles from SQLAlchemy backend
